Introduce Translator singleton.
