+++ openslides_backend/locale/en_US/LC_MESSAGES/messages.po
@@ -31,5 +31,5 @@ msgstr ""
 
 #: action/actions/user/forget_password.py:28
 msgid "Reset your OpenSlides password"
-msgstr "blablabla"
+msgstr ""
